refactor: move some dataio funcs into subpkg
