+++ src/student_success_tool/dataio/__init__.py
@@ -1,4 +1,6 @@
 from . import pdp
+from .assets import read_features_table
 from .config import read_config
+from .models import load_mlflow_model
 from .read import from_csv_file, from_delta_table
 from .write import to_delta_table
