+++ src/student_success_tool/modeling/utils.py
@@ -1,18 +1,10 @@
 import logging
-import pathlib
 import typing as t
 from collections.abc import Sequence
 
-import mlflow
 import numpy as np
 import pandas as pd
 import sklearn.utils
-
-try:
-    import tomllib  # noqa
-except ImportError:  # => PY3.10
-    import tomli as tomllib  # noqa
-
 
 LOGGER = logging.getLogger(__name__)
 
@@ -84,59 +76,3 @@ def compute_sample_weights(
         dtype="float32",
         name="sample_weight",
     )
-
-
-def load_features_table(fpath: str) -> dict[str, dict[str, str]]:
-    """
-    Load a features table mapping columns to readable names and (optionally) descriptions
-    from a TOML file located at ``fpath``, which can either refer to a relative path in this
-    package or an absolute path loaded from local disk.
-
-    Args:
-        fpath: Path to features table TOML file relative to package root or absolute;
-            for example: "assets/pdp/features_table.toml" or "/path/to/features_table.toml".
-    """
-    pkg_root_dir = next(
-        p
-        for p in pathlib.Path(__file__).parents
-        if p.parts[-1] == "student_success_tool"
-    )
-    file_path = (
-        pathlib.Path(fpath)
-        if pathlib.Path(fpath).is_absolute()
-        else pkg_root_dir / fpath
-    )
-    with file_path.open(mode="rb") as f:
-        features_table = tomllib.load(f)
-    LOGGER.info("loaded features table from '%s'", file_path)
-    assert isinstance(features_table, dict)  # type guard
-    return features_table
-
-
-def load_mlflow_model(
-    model_uri: str,
-    framework: t.Optional[t.Literal["sklearn", "xgboost", "lightgbm"]] = None,
-) -> object:
-    """
-    Load a (registered) MLFlow model of whichever model type from a specified URI.
-
-    Args:
-        model_uri
-        framework
-
-    References:
-        - https://mlflow.org/docs/latest/python_api/mlflow.sklearn.html#mlflow.sklearn.load_model
-        - https://mlflow.org/docs/latest/python_api/mlflow.pyfunc.html#mlflow.pyfunc.load_model
-    """
-    load_model_func = (
-        mlflow.sklearn.load_model
-        if framework == "sklearn"
-        else mlflow.xgboost.load_model
-        if framework == "xgboost"
-        else mlflow.lightgbm.load_model
-        if framework == "lightgbm"
-        else mlflow.pyfunc.load_model
-    )
-    model = load_model_func(model_uri)
-    LOGGER.info("mlflow model loaded from '%s'", model_uri)
-    return model
